+++ local.js
@@ -1,2 +1,2 @@
-const getFromSpread = require('./index.js').getFromSpread;
+const getFromSpread = require('./src/index.js').getFromSpread;
 getFromSpread({dataType: 'kodomoSyokudo'}, (_, value) => console.log(value) );
